load user to state on login and add navigation to Profile or FirstLogin
